+++ src/tt.cpp
@@ -207,4 +207,4 @@ TTEntry::TTEntry() {
 TTEntry::TTEntry(Key k, Value v, ValueType t, Depth d, Move m,
                  int generation) :
   key_ (k), data((m & 0x7FFFF) | (t << 20) | (generation << 23)),
-  value_(v), depth_(int16_t(d)) {}
+  value_(int16_t(v)), depth_(int16_t(d)) {}
